Converted amounts in expenses and transactions from floats to BigDecimal and made fraction length depend on currency

+++ client/src/main/java/client/scenes/MainCtrl.java
@@ -368,7 +368,7 @@ public class MainCtrl implements MainCtrlInterface{
         addExpenseCtrl.setExpenseAuthor(exp.getExpenseAuthor().getName());
         addExpenseCtrl.setPurpose(exp.getPurpose());
 
-        addExpenseCtrl.setAmount(exp.getAmount(), exp.getDate(), exp.getCurrency());
+        addExpenseCtrl.setAmount(exp.getAmount().doubleValue(), exp.getDate(), exp.getCurrency());
         addExpenseCtrl.setCurrency(exp.getCurrency());
         addExpenseCtrl.setDate(exp.getDate().toInstant().
                 atZone(ZoneId.systemDefault()).toLocalDate());
